requirements.txt
altair==4.2.0
altgraph==0.17.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==41.0.3
cx-Freeze==6.15.5
cx-Logging==3.1.0
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-allauth==0.55.0
django-autoslug==1.9.8
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tinymce==3.5.0
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.1
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.3
kiwisolver==1.4.4
lief==0.13.2
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.2
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.2.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.8.2
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.6.2
streamlit==1.17.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
virtualenv==20.17.1
watchdog==2.2.1
wcwidth==0.2.6
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
